requirements.txt
mysql-connector-python==8.0.26
elasticsearch==7.17.0
Flask==2.0.1
Werkzeug==2.0.3
python-dotenv==0.19.0